Latest changes

+++ src/client/js/app/filter/main.js
@@ -5,61 +5,80 @@
 (function(filter) {
     'use strict';
 
-    filter.filterNumber = 0;
+    var current;
 
     // On start, bind metrics
     filter.start = function() {
         filter.color();
-        filter.entities();
-        filter.keywords();
-
         $('.tooltip').tooltipster();
+        filter.initSelector();
 
-        $('[data-bind]').on('click', function(e){
+        $('#filter-estimate').on('click', function(e) {
             e.preventDefault();
-            hunk.filter[$(this).attr('data-bind')]();
-        });
-
-        if ($('[name=keywords]').val() != '') {
             filter.estimate();
+        });
+        $('#filter-submit').on('click', function(e) {
+            e.preventDefault();
+            filter.submit();
+        });
+        $('#filter-remove').on('click', function(e) {
+            e.preventDefault();
+            filter.remove();
+        });
+    };
+
+    filter.init = function(f) {
+        var i, l, conds;
+
+        if (!!f) {
+            current = f;
+        } else {
+            current = {
+                name: '',
+                color: '#333333',
+                conditions: []
+            }
         }
+
+        console.log(current);
+        conds = filter.fromDB(current);
+
+        for (i = 0, l = conds.length; i < l; i += 1) {
+            if (!!filter.views[conds[i].id]) {
+                filter.views[conds[i].id](conds[i]);
+            }
+        }
+
     };
 
     filter.gather = function() {
         var f = {
             name: $('[name=name]').val(),
             color: $('[name=color]').val(),
-            entities: $('[name=entities]').val(),
-            keywords: $('[name=keywords]').val(),
-        },
-        i = 0,
-        key;
-
-        while(true) {
-            key = $('[name=key\\[' + i + '\\]]');
-            if (key.length === 0) {
-                break;
-            }
-
-            f[key.val()] = $('[name=val\\[' + i + '\\]]').val();
-
-            i += 1;
-        }
+            conditions: filter.toDB()
+        };
 
         return f;
     };
 
-    filter.remove = function() {
-        $('#estimation').html(hunk.tpl.render('tpl-loading-state'));
 
-        $.ajax({
-            method: 'DELETE',
-            contentType: 'application/json',
-            url: $('#filter-remove').attr('href'),
-        })
-        .done(function(data) {
-            console.log(data);
-        });
+
+
+
+
+    filter.remove = function() {
+        if (window.confirm('Are you sure?')) {
+            $('#estimation').html(hunk.tpl.render('tpl-loading-state'));
+
+            $.ajax({
+                method: 'DELETE',
+                contentType: 'application/json',
+                url: $('#filter-remove').attr('href'),
+            })
+            .done(function(data) {
+                console.log(data);
+            });
+        }
     };
 
     filter.submit = function() {
@@ -69,7 +88,7 @@
             method: 'POST',
             data: JSON.stringify(filter.gather()),
             contentType: 'application/json',
-            url: $('#filter-form').attr('data-url'),
+            url: $('.filter-settings').attr('data-url'),
         })
         .done(function(data) {
             console.log(data);
@@ -77,8 +96,6 @@
     };
 
     filter.estimate = function() {
-        $('#estimation').html(hunk.tpl.render('tpl-loading-state'));
-
         $.ajax({
             method: 'POST',
             data: JSON.stringify(filter.gather()),
